faq-automator/requirements.txt
# requirements.txt

# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# AI/ML
google-generativeai==0.3.2
sentence-transformers==2.3.1
langchain==0.1.16
langchain-text-splitters==0.0.1
langgraph==0.0.26
faiss-cpu==1.7.4
faster-whisper==0.10.0

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3

# Twilio
twilio==8.11.0

# Firebase
firebase-admin==6.3.0

# Utilities
python-dotenv==1.0.0
aiohttp==3.9.1
requests==2.31.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.26.0